refactor: wallet 패키지 분리
