--- src/main/java/com/example/simplepayment/controller/request/CreateWalletRequest.java
@@ -1,4 +1,0 @@
-package com.example.simplepayment.controller.request;
-
-public record CreateWalletRequest(Long userId) {
-}
