+++ src/main/java/com/example/simplepayment/wallet/domain/WalletRepository.java
@@ -1,4 +1,4 @@
-package com.example.simplepayment.domain;
+package com.example.simplepayment.wallet.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
