+++ src/main/java/com/example/simplepayment/wallet/presentation/request/CreateWalletRequest.java
@@ -1,0 +1,4 @@
+package com.example.simplepayment.wallet.presentation.request;
+
+public record CreateWalletRequest(Long userId) {
+}
