+++ src/test/groovy/com/example/simplepayment/wallet/presentation/WalletControllerTest.groovy
@@ -1,11 +1,12 @@
-package com.example.simplepayment.controller
+package com.example.simplepayment.wallet.presentation
 
-import com.example.simplepayment.controller.request.AddBalanceWalletRequest
-import com.example.simplepayment.controller.request.CreateWalletRequest
-import com.example.simplepayment.controller.response.AddBalanceWalletResponse
-import com.example.simplepayment.controller.response.CreateWalletResponse
-import com.example.simplepayment.controller.response.SearchWalletResponse
-import com.example.simplepayment.service.WalletService
+
+import com.example.simplepayment.wallet.presentation.request.AddBalanceWalletRequest
+import com.example.simplepayment.wallet.presentation.request.CreateWalletRequest
+import com.example.simplepayment.wallet.presentation.response.AddBalanceWalletResponse
+import com.example.simplepayment.wallet.presentation.response.CreateWalletResponse
+import com.example.simplepayment.wallet.presentation.response.SearchWalletResponse
+import com.example.simplepayment.wallet.application.WalletService
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.spockframework.spring.SpringBean
 import org.springframework.beans.factory.annotation.Autowired
